separated main button components
