+++ src/Pomodoro.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
-import classNames from "./utils/class-names";
 import useInterval from "./utils/useInterval";
 import Timer from "./Timer";
 import SessionInfo from "./SessionInfo";
+import PlayStopButtons from "./PlayStopButtons";
 function Pomodoro() {
   // STATES
   //set focus
@@ -21,54 +21,16 @@ function Pomodoro() {
   useInterval(
     () => {
       // ToDo: Implement what should happen when the timer is running
-      //onPlay();
+      onPlay();
     },
     isTimerRunning ? 1000 : null
   );
-
-  // On Play/Pause
-  function PlayPause() {
-    // toggle timer
-    setIsTimerRunning((prevState) => !prevState);
-    if (focusCount === focusDuration * 60 && isTimerRunning === false) {
-      setFocusCount(focusDuration * 60);
-    }
-    if (breakCount === breakDuration * 60 && isTimerRunning === false) {
-      setBreakCount(breakDuration * 60);
-    }
-    if (isTimerRunning === true) {
-      console.log("PAUSE");
-      disableTimerBtns();
-    } else if (isTimerRunning === false) {
-      console.log("PLAY");
-      enableStopButton();
-      disableTimerBtns();
-    }
-  }
 
   // ON PLAY - runs every second
   const onPlay = () => {
     runFullSession();
     if (focusCount === 0 && breakCount === breakDuration * 60) handleAudio();
     if (breakCount === 0) handleAudio();
-  };
-
-  // ON STOP
-  const onStop = () => {
-    console.log("Stop");
-    setIsTimerRunning(false);
-    setIsFocus(true);
-    setFocusCount(focusDuration * 60);
-    setBreakCount(breakDuration * 60);
-    disableStopButton();
-    enableTimerButtons();
-    // set session info
-    const sessionInfo = document.querySelector("#sessionInfo");
-    sessionInfo.classList.add("d-none");
-    sessionInfo.classList.remove("d-block");
-    const progressBar = document.querySelector(".progress");
-    progressBar.classList.add("d-none");
-    progressBar.classList.remove("flex");
   };
 
   // Session Handler
@@ -105,31 +67,6 @@ function Pomodoro() {
   // Helpers
   const audioElement = document.getElementsByClassName("audio")[0];
 
-  function enableStopButton() {
-    const stopBtn = document.querySelector("#stopBtn");
-    stopBtn.classList.remove("disabled");
-  }
-
-  function disableStopButton() {
-    const stopBtn = document.querySelector("#stopBtn");
-    stopBtn.classList.add("disabled");
-    stopBtn.getAttribute("disabled", true);
-  }
-
-  function enableTimerButtons() {
-    const timerBtns = document.querySelectorAll("#timerButton");
-    for (let btn of timerBtns) {
-      btn.removeAttribute("disabled");
-    }
-  }
-
-  function disableTimerBtns() {
-    const timerBtns = document.querySelectorAll("#timerButton");
-    for (let btn of timerBtns) {
-      btn.setAttribute("disabled", true);
-    }
-  }
-
   function handleAudio() {
     audioElement.play();
   }
@@ -147,15 +84,17 @@ function Pomodoro() {
     <div className="container">
       <div className="pomodoro">
         <div className="row">
+          {/* FOCUS TIMER */}
           <Timer
             title="Focus Duration"
             upperLimit={60}
-            lowerLimit={5}
-            changeInterval={5}
+            lowerLimit={1}
+            changeInterval={1}
             value={focusDuration}
             onChange={setFocusDuration}
             disabled={isTimerRunning}
           />
+          {/* BREAK TIMER */}
           <Timer
             title="Break Duration"
             upperLimit={15}
@@ -166,56 +105,28 @@ function Pomodoro() {
             disabled={isTimerRunning}
           />
         </div>
-        <div className="row">
-          <div className="col">
-            <div
-              className="btn-group btn-group-lg mb-2"
-              role="group"
-              aria-label="Timer controls"
-            >
-              <button
-                type="button"
-                className="btn btn-primary"
-                data-testid="play-pause"
-                title="Start or pause timer"
-                onClick={PlayPause}
-              >
-                <span
-                  className={classNames({
-                    oi: true,
-                    "oi-media-play": !isTimerRunning,
-                    "oi-media-pause": isTimerRunning,
-                  })}
-                />
-              </button>
-              {/* TODO: Implement stopping the current focus or break session and disable when there is no active session */}
-              <button
-                type="button"
-                className="btn btn-secondary disabled"
-                id="stopBtn"
-                title="Stop the session"
-                onClick={onStop}
-              >
-                <span className="oi oi-media-stop" />
-              </button>
-            </div>
-          </div>
-        </div>
-        <div>
-          {/* TODO: This area should show only when a focus or break session is running or pauses */}
-          <div className="row mb-2 d-none" id="sessionInfo">
-            <SessionInfo
-              isFocus={isFocus}
-              focusDuration={focusDuration}
-              breakDuration={breakDuration}
-              focusCount={focusCount}
-              breakCount={breakCount}
-              isTimerRunning={isTimerRunning}
-              barWidth={barWidth}
-              ariaValue={ariaValue}
-            />
-          </div>
-        </div>
+        <PlayStopButtons
+          isTimerRunning={isTimerRunning}
+          setIsTimerRunning={setIsTimerRunning}
+          setIsFocus={setIsFocus}
+          setFocusCount={setFocusCount}
+          focusDuration={focusDuration}
+          setBreakCount={setBreakCount}
+          breakDuration={breakDuration}
+          focusCount={focusCount}
+          breakCount={breakCount}
+        />
+        {/* shows only during active focus or break session */}
+        <SessionInfo
+          isFocus={isFocus}
+          focusDuration={focusDuration}
+          breakDuration={breakDuration}
+          focusCount={focusCount}
+          breakCount={breakCount}
+          isTimerRunning={isTimerRunning}
+          barWidth={barWidth}
+          ariaValue={ariaValue}
+        />
       </div>
     </div>
   );
